[FIX] 회원 수정 오류 해결, 강아지 생일 응답 (#58)
* [FEAT] 리프레시 토큰 추가, 로그인시 로그인 데이터 반환

* [ADD] 회원 역할 enum타입 변경, 예외처리

* [FIX] 회원 수정 오류 해결, 강아지 생일 응답

+++ server/src/main/java/com/main026/walking/community/service/CommunityService.java
@@ -5,6 +5,8 @@ import com.main026.walking.community.dto.CommunityDto;
 import com.main026.walking.community.entity.Community;
 import com.main026.walking.community.mapper.CommunityMapper;
 import com.main026.walking.community.repository.CommunityRepository;
+import com.main026.walking.exception.BusinessLogicException;
+import com.main026.walking.exception.ExceptionCode;
 import com.main026.walking.image.entity.Image;
 import com.main026.walking.image.repository.ImageRepository;
 import com.main026.walking.member.entity.Member;
@@ -20,6 +22,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -43,8 +46,6 @@ public class CommunityService {
     //  Create
     public Community createCommunity(CommunityDto.Post postDto,Member member) {
         Community community = communityMapper.postDtoToEntity(postDto);
-        //TODO 임시방편 : 매핑과정에서 조회수와 좋아요의 기본값이 설정되지않고있다.
-        community.setViewAndLike();
 
         String[] dayInfo = postDto.getDates();
         List<String> dayList = new ArrayList<>();
@@ -75,7 +76,6 @@ public class CommunityService {
         } catch (IOException e) {
             e.printStackTrace();
         }
-        System.out.println(community.getViewed());
         return communityRepository.save(community);
     }
 
@@ -116,17 +116,21 @@ public class CommunityService {
     }
 
     //  Delete
-    public void deleteCommunity(long communityId) {
+    public void deleteCommunity(long communityId, PrincipalDetails principalDetails) {
+        Long authId = principalDetails.getMember().getId();
+        Long representId = communityRepository.findById(communityId).orElseThrow().getRepresentMember().getId();
+        if (authId!=representId){
+            throw new BusinessLogicException(ExceptionCode.NO_AUTHORIZATION);
+        }
         Community community = findVerifiedCommunity(communityId);
         communityRepository.delete(community);
     }
 
     //  Valid
     public Community findVerifiedCommunity(long questionId) {
-        Optional<Community> optionalCommunity =
-                communityRepository.findById(questionId);
+        Optional<Community> checkCommunity = communityRepository.findById(questionId);
         Community findCommunity =
-                optionalCommunity.orElseThrow(() -> new RuntimeException("COMMUNITY_NOT_FOUND"));
+                checkCommunity.orElseThrow(() -> new RuntimeException("COMMUNITY_NOT_FOUND"));
 
         return findCommunity;
     }
